After some discussion, (parts of) the KDE FreeBSD team have picked k3b as "most likely to be portable". At least it looks ok. This means investing some time and energy (on our part) in identifying all the Linuxisms, #ifdeffing them, and adding equivalent functionality. It's time for the age of "KDE supports Linux only for CD burning" to come to an end.
This first commit just makes k3bmmc.h compile by dealing with
the different #includes for endianness. I actually don't think
this is the best way of #ifdeffing, since Linux OS-specific
defines are a mess. Perhaps adding a k3bsystem.h that uses the
Qt logic to identify the OS would be useful -- and make it bomb
out on unsupported systems. Then we could rely on #ifdef Q_OS_LINUX
xor Q_OS_FREEBSD xor ...

If this fails to compile for you on Linux, check the preprocessor
define that your system happens to use, with something like
	gcc -dumpspecs | grep -i bsd
and then add that to the #ifdef linux line, or consider Q_OS_LINUX.

CCMAIL: kde@freebsd.org

svn path=/trunk/kdeextragear-1/k3b/; revision=224182