Well, the audio output stuff sucked big time so I rewrote it completely. It's now based on K3bAudioOutputPlugins which are much simpler and easier to handle. It's not perfect but much better than before. For now we have an arts plugin. Alsa would be nice. :) It can be configured in the settings dialog or choosen via command line argument --ao (ie. "k3b --ao arts")
Added an animated pixmap to the audio project indicating th currently playing track
(inspired by amarok).

Added a play action to the audio project to play the selected track.

Cleaned up the docking stuff. Now the project view cannot be hidden anymore and none of
the dock windows can be docked to the desktop (made a floating window). Finally!

svn path=/trunk/kdeextragear-1/k3b/; revision=369057