Created proper doxygen documentation for libk3bdevice.
Created a new plugin type "projectplugin" and converted the audio cddb plugin
and the meta info renamer plugin to the new interface. This interface is not
perfect at all so some input would be nice.

Renamed all the theme pixmaps to give them proper names.

Added a toolbar for every project.

svn path=/trunk/kdeextragear-1/k3b/; revision=390537