I finally did the big libk3b cleanup. It is almost done now. That means that libk3b will be usable for other applications soon (like a K3bLite for example).
Added MusicBrainz support.

Added a better audio file conversion based on the audio project. You get the
same dialog like for Audio CD ripping. That means it is now possible to convert
an audio project in any wanted format and create a playlist from it or even a cue
file.

svn path=/trunk/kdeextragear-1/k3b/; revision=391594