Automatically setect "only-create-image" if no burner is found.
Removed the confusing "list hidden" and "list system" options and
replaced them with a messagebox which asks if the user wants to add
those files (I think it's way more user-friendly)

Fixed a crash in K3bDataFileView when dropping files after the selected
dir has been removed.

Moved the "ask to save project" option to the misc tab.

Grab focus after dropping files so actions like delete will happen in
the project and not in the file browser.

svn path=/trunk/extragear/multimedia/k3b/; revision=418644