For all you kioslave addicts:
This one does VideoDVD ripping. Or better: It simply allows you to copy
all the files from the VideoDVD (that means the vobs and stuff) with
on-the-fly decryption (if you are so lucky to live in a country that does
allow to install libdvdcss ;).
Yep, it's as simple as typing
videodvd:/
in the Konqueror adress field and simply copy like you always do. What you get 
is a decrypted copy of your VideoDVD.

The coolest thing is that it only took me one afternoon to hack this together
since the k3blibs offer all the stuff I needed.

Have fun. :)

CCMAIL: kde-extra-gear@kde.org
CCMAIL: k3b-users@lists.sourceforge.net

svn path=/trunk/extragear/multimedia/k3b/; revision=423965