New DCOP interface:  K3bJobInterface
The K3bJobInterface provides DCOP signal equivalents to all the K3bJob signals
which emit the state of the currently running job in K3b (for example the progress(int)
signal).
Using this interface it is possible to monitor the progress of a job from another application
or script. For testing I tried the dcoppython bindings which worked nicely.

This may not be a very important feature but it was fun to implement and only took half an hour.
Maybe someday someone wants to create a Karamba theme or something like that...

svn path=/trunk/extragear/multimedia/k3b/; revision=511965