cleanups for more KProcess porting
This introduces three new methods of K3bProcess that work similar to those in 
QProcess and have the same names. They are in fact limited to the currently 
used functionality.

This reduced the amount of files that need to be touched on the final step and 
also stops classes to do nasty things directly on the descriptors of K3bProcess.

svn path=/trunk/extragear/multimedia/k3b/; revision=824287