libk3b: Use QLibrary instead of dlopen directly.
This commit makes the code much shorter by delegating dlopen and its
related calls to QLibrary.

It thus removes much of the code introduced in b7cccda4 to make K3b
build on Windows.

Additionally, the libraries are unloaded in some more error conditions,
and we do not hardcode the major library number for libdvdcss anymore,
as that varies across Linux and FreeBSD, for example.

REVIEW: 101038