Prefer growisofs to wodim for DVD/BluRay burning.
K3b 2 defaults to cdrecord for all burning tasks, including DVDs and BluRay
discs. Unfortunately, it also does this when cdrecord is actually wodim. This
is a bad idea, because wodim's DVD burning code is not the "ProDVD" code in
Jörg Schilling's current cdrecord releases, but a much older, buggier and
basically unmaintained DVD patch. We cannot ship the ProDVD code in wodim
because of licensing conflicts: That code was never released under the GPL, it
was relicensed directly from its original proprietary license to the CDDL. But
wodim is GPLed, and cannot be relicensed to the CDDL, in fact this was the
whole reason for the fork: Jörg Schilling's cdrecord distributes mixed CDDL and
GPL code linked together. So the DVD code in wodim is based on an ancient
experimental community-contributed DVD support patch for cdrecord (from the
times where ProDVD was entirely proprietary). So it's a bad idea to use wodim
for DVDs. As for BluRay discs, those aren't currently supported by wodim at
all; K3b should detect this, but still, it's better to explicitly default to
growisofs there too, in case wodim grows some experimental BluRay support.

One concrete known issue with wodim's DVD burning code is that it fails to burn
dual-layer DVD+Rs: https://bugzilla.redhat.com/show_bug.cgi?id=610976 . But
chances are there are many more DVD burning bugs in wodim, which are unlikely
to get fixed promptly.

Growisofs, on the other hand, is designed specifically for DVDs and BluRay
disks, doesn't have licensing issues and has been used successfully for DVDs
for years (in fact, K3b 1 always used growisofs for DVDs).

This patch makes K3b default to growisofs for all DVD or BluRay burning tasks
if cdrecord is actually wodim.

REVIEW: 101208