Removed dialog asking what set of settings should be loaded at next burn.
Most of the new users are not able to answer this question sensibly. Not because they're dumb but because they don't know the application well enough and it's hard for them to forsee the consequences of their choice. For this reason the dialog is frustrating.
Now by default the saved settings will be loaded which is IMHO the best choice for most users.
BUG: 196604