Fixed crash on detecting writing speeds
In method K3b::Device::Device::modeSense() page data is initially
resized to 0xFFFF, the page date is being read and then the actual
size of the page is obtained from the data itself. The problem was,
after reading the actual size, data was not resized back to it.
Additionally, Device::Device::determineSupportedWriteSpeeds() method
lacked checking if data array is of enough size before accessing it.

BUG: 272427
FIXED-IN: 2.0.3