Decoupled DataProjectModel from DataUrlAddingDialog
DataProjectModel no longer depends on DataUrlAddingDialog. Previously
when drag/drop action was performed DataUrlAddingDialog handled it.
Now, addUrlsRequested()/moveItemsRequested() are called which are
connected to their counterparts in DataUrlAddingDialog in the higher
layers.
This change was added to allow testing of DataProjecModel without
pulling the half of K3b with it.