Inform the user if the inserted medium is not suitable
In the burn dialog, the message displayed when an incorrect medium was
inserted was the same that the one when there was no medium inserted.
As a result, the user might think that there's a problem with the device
instead of a problem with the medium.

If there's a medium inserted on a device but it's no suitable (for
example if it's a CD and it should be a DVD, or if it's a read-only
medium, or if the capacity of the medium is not enough), the user will
be informed of the reason of the issue.

mediaRequestStrings added as a static method to Medium class. This
method takes as parameters a list of unsuitable mediums and the
requeriments for a medium to be suitable and returns a QStringList with
messages describing the problem of each unsuitable medium.

MediaSelectionComboBox was modified to call mediaRequestStrings when
a medium is inserted but is not suitable.

BUG: 259761
FIXED-IN: 2.1.0
REVIEW: 110230