cmake: use a search module for iconv
Import a FindIconv.cmake (from libstreams) and use it to look for iconv
availability, instead of just checking for iconv.h; this helps in
checking whether there is a library needed, and also whether "const"
is needed in the 2nd arg of iconv().

Use the FindIconv module, and set HAVE_ICONV (renamed from HAVE_ICONV_H)
properly.