Fix path traversal in Album Artist field for CD ripping
For the ALBUMARTIST case in k3bpatternparser, the "/" character
was not being replaced with "_" the way it was for all the other
string fields.  This enables path traversal if the CDDB data has a
"/" in the Album Artist field, creating unwanted extra directories.
This patch fixes it by performing the same replacements on the
album artist as for all the other string fields.

(cherry picked from commit 3e20e668c6137ec85a0d16d537f2c88ea40ef654)