Settings (External Programs): Give a useful error message on failure
If the authorisation dialogue fails or is cancelled, KJob::error()
is nonzero but KJob::errorString() is blank.  In this case, it is
necessary to decode the KAuth::Action::AuthStatus to obtain a readable
error message.  New function K3b::ExternalBinWidget::authErrorString()
implemented as class-static because it may be useful elsewhere.

I18N: